requirements.txt
requests==2.22.0
pymongo==3.8.0
ipdb==0.12
python-dotenv==0.10.2
sentry-sdk==0.8.0
